Changed the way that t_start >= t_end is handled. Previously, BNG would throw an error and quit. Now, BNG throws a warning that the simulation won't run and then calls run_network as usual, which prints the initial concentrations and then quits normally.
This is useful, e.g., when running multiple sequential simulations that
involve logical stopping conditions. The final time for each simulation
is indeterminate so it is unclear, a priori, how many simulation runs
will be necessary to reach t_end. Currently, BNG quits with an error
once t_end is reached because subsequent simulations have t_start =
t_end. Quitting gracefully, rather than with an error, is obviously
preferable in this case.