List of initial assignments for constant expressions in writeSBML
Changed the sbml output to create a list of initial assignments when defining a paramenter with a constant expression. Previously this value would be defined through a assignment rule which made it difficult to differentiate
from true observables. Changed constant field of the parameter definition to true. Changed overall document version to Level 2 version 2. The change passes the online validation test

TODO: the size field for the compartment definition should be a double value instead of the current parameter reference

Former-commit-id: d6ed7b298f0fed58d483947c6164f9b165c7acc7