bugfix: initialAssignments and reverse dependenies. added: lexical analysis
Fixed a bug in initial assignments where it wouldnt check if a parameter had been previously initialized as zer0
Fixed a bug in the reverse dependency check (deducing the structure of simpler
elements from more complex ones) where it wasn't necessary

Added: lexical analysis in two cases. When it knows that two species are
related but it doesn't know how (for example when you have a species with multiple
modification candidates) now it tries to apply lexical analysis (can it deduce it
based only on the naming structure without regards for any reaction information)

When it doesn't have any info regarding a given species, it will also try to apply
naming conventions.


Former-commit-id: 6ec8be1e0677fbb306b9180ebbb743d77d70abde