bugfix: fixed a bug related to structure discovered by lexical analysis
Fixed a bug that was caused when the lexical analyzer identified the structure of a molecule
using 2 molecules that were not used anywhere in the system (reactions) themselves. Since those parent molecules
were not used in the system they were not being added to the dependency graph. The fix consisted in checking
for all molecules used and dicovered through lexical analysis to be properly added in the dependency tree