bugfix: incorrect indexing when adding component to binding partners
Corrected a bug when it would poll the wrong molecule index when checking for whether a component was already existing.
This was not causing a problem before since you only need a single binding
partner to identify a bond, but it was causing incosistent selection of binding partners when dealing with dimers.

This commit also restores several changes previously existing in commits b8134433ec43176bab25567c065ce8235cbff94d and
6560cc157bc5570743e552d96607049cb33e53e2 that for some reason where lost in
my local copy