accounting for species compartment when creating product pattern
This bugfix addesses the bug reported in issue #141. The cause was that
species compartment information was getting lost when generating the list of
product SpeciesGraph's from a given reactant pattern in rxnrule->apply_operations. This
information was still present in the original patterns rxnrule->products so it was a matter
of copying the information over.