improving complex modification algorithm
Improving the algorithm that is in charge of mapping complex modifications (modifciations that are the combination of two or more modifications) to a set of simpler modifications.
Created function testAgainstExistingConventions in analyzeSBML that makes use of itertools fucntions ifilter and permutations to find a suitable combination of existing modification
that can address a more complex one. Added the memoize tag to this function since it can be quite expensive if done repeatedly.