Discard bidirectionality in some cases. More context processing
Introduced two main changes:

Discard bidirectionality in reversible reactions whose rates are constant.
This comes from the fact that SBML reactions are bidirectional by default
and models sometime forget to set the reversible flag to false. This was causing
problems in context analysis

added a context analysis summary printing function to the atomizer. This
creates a new file with the _context.log suffix where it prints information
about context motifs that atomizer has deemed to be suspect, along with
associated species where it can find them.

Improved the detail of some of the log comments printed where there is a conflict
between stoichiometry and lexical candidates.