adding back biogrid querying to bond analysis + minor formatting
BioGrid is a much faster way, albeit less comprehensive, to finding
protein-protein interactions compared to pathway commons. (pathway commons
is much more comprehensive, and hence can take a longer amount of time). updated
the system so that it uses biogrid by default and pwcm as a fallback if no candate binding information
is found.