Update README.md
Format some stuff in "Troubleshooting" section.

Better explain how to find the cvode or muparser sources. I'm not sure why we were referencing the github repo. They are either reading the directions at github itself already or they are working with the source directly. In either case, just tell them where the libsource directory is in the repo.