Move git submodules out of Network3.
Jim pointed out that it probably made more sense to have the git
submodules (currently NFSim and atomizer) out of the Network3 folder.
Now, they're just in bionetgen/bng2. There is a makefile in that dir
that builds just those two things and the Network3 makefile has been
reverted back to essentially how it was before the submodules were
added. It probably makes sense for the Makefile in bng2 to call the
Network3 makefile, which I'll do in a future commit.