Fix problems with atomizer dependency.
All of the changes to atomizer right now are happening on the python3
branch, and the build system there has changed slightly. So, now, I'm
bringing those changes into the bionetgen repo, which mainly  means
updating the submodule itself and the CI systems. This commit should fix
travis, but appveyor still needs to be done.