[WIP] sundials 3 compatiblity
I found [1] and creates a WIP patch that compiles with sundials 3+.
The part for "dense" is directly based on that so it's probably at
least partially right. The "sparse" part is almost certainly wrong,
because I wasn't sure what to put in various parameters. Somebody who
understands this code better can hopefully fix that. I also don't know
how to test this code. But I think this might be useful as a starting
point.

It seems that the solver should be initialized after the check for
Jacobian (doesn't make sense to create it and then error out), so I
moved that down.

[1] http://sundials.2283335.n4.nabble.com/Usage-notes-lead-to-the-example-that-uses-the-non-existent-header-cvode-cvode-dense-h-td4654260.html