Update network.cpp
Attempt to fix bug in output of .net files with either reactants or products that are just '0' -- i.e. production or degradation reactions. When reactions are read in that just have a 0 as either reactant or product, it results in the r_index or p_index arrays having zero length, but this case was not handled correctly in this function.