First pass at removing reaction symmetry calculations.
Turns off calculation of reaction symmetry if
1) Only syn/del ops in the rule.
2) >3 reactants in rule.
The former makes it possible to translate "flat" rules without having to try to recalculate symmetry factors. Both changes avoid the case there there are a large number of identifical reactants, which can trigger a search for automorphisms that is N! This could probably be avoided by using nauty.